fix:deployment failing due to /features
