+++ data/features.ts
@@ -1,0 +1,47 @@
+import {
+	GraduationCap,
+	Trophy,
+	LineChart,
+	Users,
+	Sparkles,
+	Shield,
+} from "lucide-react";
+
+export const featuresData = [
+	{
+		Icon: GraduationCap,
+		title: "Interactive Learning",
+		description:
+			"Engaging lessons on DeFi, staking, and blockchain fundamentals with quizzes and hands-on exercises.",
+	},
+	{
+		Icon: LineChart,
+		title: "Real Yield Staking",
+		description:
+			"Stake your tokens and earn competitive APY while learning about DeFi mechanisms.",
+	},
+	{
+		Icon: Trophy,
+		title: "Learning Rewards",
+		description:
+			"Earn APY boosts and bonus tokens by completing educational modules and quizzes.",
+	},
+	{
+		Icon: Shield,
+		title: "Secure Platform",
+		description:
+			"Enterprise-grade security with OCID authentication and multi-layer protection.",
+	},
+	{
+		Icon: Users,
+		title: "Community Learning",
+		description:
+			"Connect with other students and educators in our vibrant community forums.",
+	},
+	{
+		Icon: Sparkles,
+		title: "Educator Tools",
+		description:
+			"Comprehensive tools for educators to create courses, track student progress, and earn rewards.",
+	},
+];
